Prototype TH13.5 Support
This adds prototype support for TH13.5. There definitely are
some optimizations that can be made, but it is working and you
can extract TH13.5 music with this.

This serves as a base for the rest of the Tasofro games that
use the same engine, namely TH14.5 and TH15.5.

Portions of the code is derived from https://github.com/brliron/135tk/
namely the th135arc-alt project within for its unpacking
implementation. Big shoutouts to brliron for said project.

With this, an additional dependency is needed, namely MIRACL for
its RSA decryption. I have added in a fork as a submodule with
project files and other necessary modifications for ease of build.