Refactor File Decryption for Music Tracks
Previously the packmethods decrypted music files in the archives
through a DecryptFile method that has no track info as a parameter.
This became a problem for TH13.5-TH17.5 where each track file had
its own key for decryption, requiring a hack.

This refactors the DecryptFile method to a DecryptMusic, taking in
a pointer to the wanted TrackInfo, rather than generic position and
size parameters, which not only allows it to get the key from the
TrackInfo but also potentially future proof it.

This does make it so that it assumes decryption won't change the size
of the file. However this assumption is present elsewhere in the code
and can potentially be dealt with in TrackInfo as well.