Refactor Dictionary with Specialized HashMap
The previous implementation used FXDict, which is a string to void*
mapping. Because of this, since we wanted the key to be the filename
hash, we did a format string to generate the key.

This new implementation uses a different hashmap that allows you
to specify the type of the key, avoiding the need to do a format string.
That said, from testing it appears it doesn't increase performance.
And it does add another library, even if it is small. So for now it
will just be in this branch