Refactor handling multiple bgmfiles for th135-175
Have DiskFN return the correct data file associated with the given
track so that we don't have to check during track decryption and
potentially temporarily open another file during that.

Ideally we probably can have this info in the bgminfo files but
I'm not that confident in the codebase yet to do that refactor and
obsoleting the BGMFile in the GameInfo structure for those games.